final-text-extractor/requirements.txt
langchain-text-splitters
langchain-community
langgraph
langchain-ollama
faiss-cpu
pandas 
markdown